GML: detect potential int overflow when reading extremely large (and hostile) files
git-svn-id: https://svn.osgeo.org/gdal/trunk@33182 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/gml/gmlreaderp.h
@@ -114,16 +114,16 @@ typedef enum
 class GMLHandler
 {
     char      *m_pszCurField;
-    size_t     m_nCurFieldAlloc;
-    size_t     m_nCurFieldLen;
+    unsigned int m_nCurFieldAlloc;
+    unsigned int m_nCurFieldLen;
     bool       m_bInCurField;
     int        m_nAttributeIndex;
     int        m_nAttributeDepth;
 
 
     char      *m_pszGeometry;
-    int        m_nGeomAlloc;
-    int        m_nGeomLen;
+    unsigned int m_nGeomAlloc;
+    unsigned int m_nGeomLen;
     int        m_nGeometryDepth;
     bool       m_bAlreadyFoundGeometry;
     int        m_nGeometryPropertyIndex;
